refactor: formatDate 함수를 유틸 함수로 분리
